Refactor findDataGraphProject method because its functionallity could be resolved by streams.
So replace that method with a new one, its name is countTaskGroupByStatus with a Project parameter,
for that new method we need to create some another methods like findTaskByProject.

I have to rename methods in the controller because there were methods with the same
name but with differents parameters, the same case with their tests.